+++ src/exceptions/ValuePHPJSError.mjs
@@ -5,6 +5,8 @@ import PHPJSError from './PHPJSError.mjs';
  * Например, передача отрицательного целого числа, когда функция ожидает положительное, или передача
  * пустой строки или массива, когда функция ожидает, что строка или массив не будет пустым.
  */
-class ValuePHPJSError extends PHPJSError {}
+class ValuePHPJSError extends PHPJSError {
+    name = 'ValuePHPJSError';
+}
 
 export default ValuePHPJSError;