+++ tests/array.global.test.mjs
@@ -44,4 +44,5 @@ test('array.global', () => {
     expect(array_walk_recursive).not.toBeUndefined();
     expect(in_array).not.toBeUndefined();
     expect(key_exists).not.toBeUndefined();
+    expect(range).not.toBeUndefined();
 });